+++ guillotina/renderers.py
@@ -164,10 +164,7 @@ class RendererRaw(Renderer):
 
     def guess_response(self, value):
         resp = value.response
-        if isinstance(resp, dict):
-            resp = aioResponse(body=bytes(json.dumps(resp, cls=GuillotinaJSONEncoder), 'utf-8'))
-            resp.headers['Content-Type'] = 'application/json'
-        elif isinstance(resp, list):
+        if type(resp) in (dict, list, int, float):
             resp = aioResponse(body=bytes(json.dumps(resp, cls=GuillotinaJSONEncoder), 'utf-8'))
             resp.headers['Content-Type'] = 'application/json'
         elif isinstance(resp, str):